Move IConfig to models package, since... it's a model :D
