+++ src/models/IConfig.ts
