+++ src/services/JWTService.ts
@@ -1,7 +1,7 @@
 import * as JWT from "jsonwebtoken";
 import AuthorizationError from "../models/exceptions/AuthorizationError";
 import {ALLOWEDROLES, ERRORMESSAGES} from "../assets/enum";
-import IConfig from "../utils/IConfig";
+import IConfig from "../models/IConfig";
 import {getCertificateChain} from "./azure";
 
 class JWTService {
